phase 3 week 23

+++ src/main/java/io/github/talelin/latticy/mapper/SpuMapper.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface SpuMapper extends BaseMapper<SpuDO> {
-    SpuDetailDO getDetail(Long id);
+    SpuDetailDO getDetail(Integer id);
 }
